\name{utilitree-internal}
\alias{bisse.ani}
\alias{consensusBrlen}
\alias{edgeLengthTaxset}
\alias{find.cherries}
\alias{heritableTransforms}
\alias{independent.pairs}
\alias{isClade}
\alias{overlayPlot}
\alias{prettyTree}
\alias{resolveNames}
\alias{saveBisseMovie}
\alias{showBisseMovie}
\alias{summarizeNode}
\title{
Undocumented functions
}
\description{
Weird (because undocumented) and wonderful (because otherwise I wouldn't have written it) things
}
\details{
Code of the functions themselves may have some comments
}
\author{
Brian O'Meara
}
